+++ runtime/development/src/migrations.rs
@@ -10,29 +10,4 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
-use cfg_primitives::AccountId;
-use sp_core::parameter_types;
-use sp_std::{vec, vec::Vec};
-
-parameter_types! {
-	// Alice
-	pub InitialTcMembers: Vec<AccountId> = vec![AccountId::new(hex_literal::hex!("d43593c715fdd31c61141abd04a99fd6822c8558854ccde39a5684e7a56da27d"))];
-}
-
-/// The migration set for Development & Demo.
-/// It includes all the migrations that have to be applied on that chain.
-pub type UpgradeDevelopment1300 = (
-	// Initialize OpenGov Technical Committee with Alice
-	runtime_common::migrations::technical_comittee::InitMigration<crate::Runtime, InitialTcMembers>,
-	runtime_common::migrations::increase_storage_version::Migration<crate::Referenda, 0, 1>,
-	runtime_common::migrations::increase_storage_version::Migration<
-		crate::TechnicalCommittee,
-		0,
-		4,
-	>,
-	runtime_common::migrations::increase_storage_version::Migration<
-		crate::TechnicalCommitteeMembership,
-		0,
-		4,
-	>,
-);
+pub type UpgradeDevelopment1301 = ();
