+++ runtime/integration-tests/src/cases/lp/investments.rs
@@ -18,8 +18,7 @@ use sp_runtime::traits::Zero;
 
 use crate::{
 	cases::lp::{
-		self, names, setup, setup_currencies, setup_deploy_lps, setup_full,
-		setup_investment_currencies, setup_pools, setup_tranches,
+		self, names, setup_full,
 		utils::{pool_c_tranche_1_id, Decoder},
 		DECIMALS_6, POOL_C,
 	},
@@ -31,7 +30,7 @@ use crate::{
 const DEFAULT_INVESTMENT_AMOUNT: Balance = 100 * DECIMALS_6;
 
 mod utils {
-	use cfg_primitives::{AccountId, Balance};
+	use cfg_primitives::AccountId;
 	use cfg_traits::{investments::TrancheCurrency, HasLocalAssetRepresentation};
 	use ethabi::Token;
 	use pallet_foreign_investments::Action;
@@ -89,7 +88,7 @@ mod utils {
 	pub fn cancel<T: Runtime>(evm: &mut impl EvmEnv<T>, who: Keyring, lp_pool: &str) {
 		evm.call(
 			who,
-			U256::zero(),
+			Default::default(),
 			lp_pool,
 			"cancelDepositRequest",
 			Some(&[Token::Uint(U256::from(0)), Token::Address(who.into())]),
@@ -156,7 +155,7 @@ mod with_pool_currency {
 	use super::{utils, *};
 	use crate::cases::lp::utils as lp_utils;
 
-	#[test_runtimes([development], ignore = "solidity mismatch")]
+	#[test_runtimes([development])]
 	fn currency_invest<T: Runtime>() {
 		let mut env = setup_full::<T>();
 		env.state_mut(|evm| {
@@ -190,7 +189,7 @@ mod with_pool_currency {
 		});
 	}
 
-	#[test_runtimes([development], ignore = "solidity mismatch")]
+	#[test_runtimes([development])]
 	fn currency_collect<T: Runtime>() {
 		let mut env = setup_full::<T>();
 		env.state_mut(|evm| {
@@ -258,7 +257,7 @@ mod with_pool_currency {
 		});
 	}
 
-	#[test_runtimes([development], ignore = "solidity mismatch")]
+	#[test_runtimes([development])]
 	fn invest_cancel_full<T: Runtime>() {
 		let mut env = setup_full::<T>();
 		env.state_mut(|evm| {
@@ -394,9 +393,11 @@ mod with_foreign_currency {
 		});
 	}
 
-	#[test_runtimes([development], ignore = "solidity mismatch")]
+	#[test_runtimes([development])]
 	fn invest_cancel_full_after_swap<T: Runtime>() {
 		let mut env = setup_full::<T>();
+
+		// Invest and swap all foreign to pool currency
 		env.state_mut(|evm| {
 			utils::invest(evm, Keyring::TrancheInvestor(1), names::POOL_A_T_1_USDC);
 			utils::fulfill_swap::<T>(
@@ -404,6 +405,7 @@ mod with_foreign_currency {
 				POOL_A,
 				pool_a_tranche_1_id::<T>(),
 				Action::Investment,
+				// Fulfill entire order
 				None,
 			);
 		});
@@ -453,22 +455,21 @@ mod with_foreign_currency {
 				None,
 			);
 
-			// todo!("@william: Use CancelDepositRequest");
-			// lp_utils::process_outbound::<T>(|msg| {
-			// 	assert_eq!(
-			// 		msg,
-			// 		MessageOf::<T>::ExecutedDecreaseInvestOrder {
-			// 			pool_id: POOL_A,
-			// 			tranche_id: pool_a_tranche_1_id::<T>(),
-			// 			investor: vec_to_fixed_array(lp::utils::remote_account_of::<T>(
-			// 				Keyring::TrancheInvestor(1)
-			// 			)),
-			// 			currency: utils::index_lp(evm, names::USDC),
-			// 			currency_payout: DEFAULT_INVESTMENT_AMOUNT,
-			// 			remaining_invest_amount: 0,
-			// 		}
-			// 	)
-			// });
+			lp_utils::process_outbound::<T>(|msg| {
+				assert_eq!(
+					msg,
+					pallet_liquidity_pools::Message::FulfilledCancelDepositRequest {
+						pool_id: POOL_A,
+						tranche_id: pool_a_tranche_1_id::<T>(),
+						investor: vec_to_fixed_array(lp::utils::remote_account_of::<T>(
+							Keyring::TrancheInvestor(1)
+						)),
+						currency: utils::index_lp(evm, names::USDC),
+						currency_payout: DEFAULT_INVESTMENT_AMOUNT,
+						fulfilled_invest_amount: DEFAULT_INVESTMENT_AMOUNT,
+					}
+				)
+			});
 
 			assert_eq!(
 				Decoder::<Balance>::decode(&evm.view(
@@ -485,7 +486,7 @@ mod with_foreign_currency {
 		});
 	}
 
-	#[test_runtimes([development], ignore = "solidity mismatch")]
+	#[test_runtimes([development])]
 	fn invest_cancel_full_after_swap_partially<T: Runtime>() {
 		let mut env = setup_full::<T>();
 		let part = Quantity::checked_from_rational(1, 2).unwrap();
@@ -557,22 +558,21 @@ mod with_foreign_currency {
 				None,
 			);
 
-			// todo!("@william: Use CancelDepositRequest");
-			// lp_utils::process_outbound::<T>(|msg| {
-			// 	assert_eq!(
-			// 		msg,
-			// 		MessageOf::<T>::ExecutedDecreaseInvestOrder {
-			// 			pool_id: POOL_A,
-			// 			tranche_id: pool_a_tranche_1_id::<T>(),
-			// 			investor: vec_to_fixed_array(lp::utils::remote_account_of::<T>(
-			// 				Keyring::TrancheInvestor(1)
-			// 			)),
-			// 			currency: utils::index_lp(evm, names::USDC),
-			// 			currency_payout: DEFAULT_INVESTMENT_AMOUNT,
-			// 			remaining_invest_amount: 0,
-			// 		}
-			// 	)
-			// });
+			lp_utils::process_outbound::<T>(|msg| {
+				assert_eq!(
+					msg,
+					pallet_liquidity_pools::Message::FulfilledCancelDepositRequest {
+						pool_id: POOL_A,
+						tranche_id: pool_a_tranche_1_id::<T>(),
+						investor: vec_to_fixed_array(lp::utils::remote_account_of::<T>(
+							Keyring::TrancheInvestor(1)
+						)),
+						currency: utils::index_lp(evm, names::USDC),
+						currency_payout: DEFAULT_INVESTMENT_AMOUNT,
+						fulfilled_invest_amount: DEFAULT_INVESTMENT_AMOUNT,
+					}
+				)
+			});
 
 			assert_eq!(
 				Decoder::<Balance>::decode(&evm.view(
@@ -679,6 +679,7 @@ mod with_foreign_currency {
 				remaining_amount
 			);
 
+			// FIXME: Fails because cannot cancel
 			utils::cancel(evm, Keyring::TrancheInvestor(1), names::POOL_A_T_1_USDC);
 
 			assert_eq!(
@@ -702,22 +703,21 @@ mod with_foreign_currency {
 				remaining_amount
 			);
 
-			// todo!("@william: Use CancelDepositRequest");
-			// lp_utils::process_outbound::<T>(|msg| {
-			// 	assert_eq!(
-			// 		msg,
-			// 		MessageOf::<T>::ExecutedDecreaseInvestOrder {
-			// 			pool_id: POOL_A,
-			// 			tranche_id: pool_a_tranche_1_id::<T>(),
-			// 			investor: vec_to_fixed_array(lp::utils::remote_account_of::<T>(
-			// 				Keyring::TrancheInvestor(1)
-			// 			)),
-			// 			currency: utils::index_lp(evm, names::USDC),
-			// 			currency_payout: remaining_amount,
-			// 			remaining_invest_amount: 0,
-			// 		}
-			// 	)
-			// });
+			lp_utils::process_outbound::<T>(|msg| {
+				assert_eq!(
+					msg,
+					pallet_liquidity_pools::Message::FulfilledCancelDepositRequest {
+						pool_id: POOL_A,
+						tranche_id: pool_a_tranche_1_id::<T>(),
+						investor: vec_to_fixed_array(lp::utils::remote_account_of::<T>(
+							Keyring::TrancheInvestor(1)
+						)),
+						currency: utils::index_lp(evm, names::USDC),
+						currency_payout: DEFAULT_INVESTMENT_AMOUNT,
+						fulfilled_invest_amount: DEFAULT_INVESTMENT_AMOUNT,
+					}
+				)
+			});
 
 			assert_eq!(
 				Decoder::<Balance>::decode(&evm.view(
