+++ runtime/integration-tests/src/cases/lp/pool_management.rs
@@ -18,7 +18,10 @@ use cfg_types::{
 	tokens::{CrossChainTransferability, CurrencyId, CustomMetadata},
 };
 use ethabi::{ethereum_types::H160, Token, Uint};
-use frame_support::{assert_ok, traits::OriginTrait};
+use frame_support::{
+	assert_ok,
+	traits::{Get, OriginTrait},
+};
 use frame_system::pallet_prelude::OriginFor;
 use pallet_evm::AddressMapping;
 use pallet_liquidity_pools::GeneralCurrencyIndexOf;
@@ -39,7 +42,7 @@ use crate::{
 	},
 };
 
-#[test_runtimes([development], ignore = "solidity mismatch")]
+#[test_runtimes([development])]
 fn add_currency<T: Runtime>() {
 	let mut env = super::setup::<T, _>(|_| {});
 
@@ -128,7 +131,7 @@ fn add_currency<T: Runtime>() {
 	});
 }
 
-#[test_runtimes([development], ignore = "solidity mismatch")]
+#[test_runtimes([development])]
 fn add_pool<T: Runtime>() {
 	let mut env = super::setup::<T, _>(|_| {});
 	const POOL: PoolId = 1;
@@ -173,7 +176,22 @@ fn add_pool<T: Runtime>() {
 	});
 }
 
-#[test_runtimes([development], ignore = "solidity mismatch")]
+#[test_runtimes([development])]
+fn hook_address<T: Runtime>() {
+	let env = super::setup::<T, _>(|_| {});
+	env.state(|evm| {
+		let solidity =
+			T::AddressMapping::into_account_id(evm.deployed(names::RESTRICTION_MANAGER).address());
+		let rust: AccountId =
+			<T as pallet_liquidity_pools::Config>::AddTrancheHookAddress::get().into();
+		assert_eq!(
+			solidity, rust,
+			"Hook address changed, please change our stored value (right) to the new address (left)"
+		);
+	})
+}
+
+#[test_runtimes([development])]
 fn add_tranche<T: Runtime>() {
 	let mut env = super::setup::<T, _>(|evm| {
 		super::setup_currencies(evm);
@@ -194,11 +212,6 @@ fn add_tranche<T: Runtime>() {
 			),
 			utils::REVERT_ERR
 		);
-
-		let hook_address = evm.deployed(names::RESTRICTION_MANAGER).address();
-		let hook_32 = T::AddressMapping::into_account_id(hook_address);
-		println!("Hook EVM Address {:?}", hook_address);
-		println!("Hook AccountId32 Address {:?}", hook_32);
 	});
 
 	env.state_mut(|_| {
@@ -260,7 +273,7 @@ fn add_tranche<T: Runtime>() {
 	});
 }
 
-#[test_runtimes([development], ignore = "solidity mismatch")]
+#[test_runtimes([development])]
 fn allow_investment_currency<T: Runtime>() {
 	let mut env = super::setup::<T, _>(|evm| {
 		super::setup_currencies(evm);
@@ -312,7 +325,7 @@ fn allow_investment_currency<T: Runtime>() {
 	});
 }
 
-#[test_runtimes([development], ignore = "solidity mismatch")]
+#[test_runtimes([development])]
 fn disallow_investment_currency<T: Runtime>() {
 	let mut env = super::setup::<T, _>(|evm| {
 		super::setup_currencies(evm);
@@ -365,7 +378,7 @@ fn disallow_investment_currency<T: Runtime>() {
 	});
 }
 
-#[test_runtimes([development], ignore = "solidity mismatch")]
+#[test_runtimes([development])]
 fn update_member<T: Runtime>() {
 	let mut env = super::setup::<T, _>(|evm| {
 		super::setup_currencies(evm);
@@ -544,7 +557,7 @@ fn update_tranche_token_metadata<T: Runtime>() {
 	});
 }
 
-#[test_runtimes([development], ignore = "solidity mismatch")]
+#[test_runtimes([development])]
 fn update_tranche_token_price<T: Runtime>() {
 	let mut env = super::setup::<T, _>(|evm| {
 		super::setup_currencies(evm);
