+++ runtime/integration-tests/src/cases/lp/transfers.rs
@@ -20,13 +20,13 @@ use frame_support::traits::OriginTrait;
 use frame_system::pallet_prelude::OriginFor;
 use pallet_liquidity_pools::Message;
 use sp_core::ByteArray;
-use sp_runtime::traits::Convert;
 
 use crate::{
 	cases::lp::{
 		self, names,
 		utils::{as_h160_32bytes, pool_a_tranche_1_id, Decoder},
-		LocalUSDC, DECIMALS_6, DEFAULT_BALANCE, EVM_DOMAIN_CHAIN_ID, POOL_A, USDC,
+		LocalUSDC, CENTRIFUGE_CHAIN_ID, DECIMALS_6, DEFAULT_BALANCE, DOMAIN_CENTRIFUGE, DOMAIN_EVM,
+		EVM_DOMAIN_CHAIN_ID, POOL_A, USDC,
 	},
 	config::Runtime,
 	env::{Blocks, Env, EnvEvmExtension, EvmEnv},
@@ -111,11 +111,23 @@ mod utils {
 				]),
 			)
 			.unwrap();
+
+			let balance = Decoder::<Balance>::decode(&evm.view(
+				Keyring::Alice,
+				names::USDC,
+				"balanceOf",
+				Some(&[Token::Address(Keyring::Alice.into())]),
+			));
+			assert!(
+				balance >= AMOUNT,
+				"Insufficient USDC funds by Alice. Required {AMOUNT} but only has {balance}"
+			);
+
 			evm.call(
 				Keyring::Alice,
 				Default::default(),
 				names::POOL_MANAGER,
-				"transfer",
+				"transferAssets",
 				Some(&[
 					Token::Address(evm.deployed(names::USDC).address()),
 					Token::FixedBytes(Keyring::Ferdie.id().to_raw_vec()),
@@ -229,6 +241,15 @@ fn transfer_tranche_tokens_domain_to_local_to_domain<T: Runtime>() {
 	utils::prepare_hold_tt_domain::<T>(&mut env);
 
 	env.state_mut(|evm| {
+		assert!(
+			Decoder::<Balance>::decode(&evm.view(
+				Keyring::TrancheInvestor(1),
+				names::POOL_A_T_1,
+				"balanceOf",
+				Some(&[Token::Address(Keyring::TrancheInvestor(1).into())]),
+			)) >= AMOUNT,
+			"Insufficient POOL_A_T_1 funds by TrancheInvestor(1)"
+		);
 		evm.call(
 			Keyring::TrancheInvestor(1),
 			Default::default(),
@@ -244,10 +265,11 @@ fn transfer_tranche_tokens_domain_to_local_to_domain<T: Runtime>() {
 			Keyring::TrancheInvestor(1),
 			sp_core::U256::zero(),
 			names::POOL_MANAGER,
-			"transferTranchesToEVM",
+			"transferTrancheTokens",
 			Some(&[
 				Token::Uint(POOL_A.into()),
 				Token::FixedBytes(pool_a_tranche_1_id::<T>().into()),
+				Token::Uint(DOMAIN_EVM.into()),
 				Token::Uint(EVM_DOMAIN_CHAIN_ID.into()),
 				Token::Address(Keyring::TrancheInvestor(2).into()),
 				Token::Uint(AMOUNT.into()),
@@ -263,14 +285,6 @@ fn transfer_tranche_tokens_domain_to_local_to_domain<T: Runtime>() {
 				Message::TransferTrancheTokens {
 					pool_id: POOL_A,
 					tranche_id: pool_a_tranche_1_id::<T>(),
-					sender:
-						<T as pallet_liquidity_pools::Config>::DomainAddressToAccountId::convert(
-							DomainAddress::evm(
-								EVM_DOMAIN_CHAIN_ID,
-								Keyring::TrancheInvestor(2).into()
-							)
-						)
-						.into(),
 					domain: Domain::EVM(EVM_DOMAIN_CHAIN_ID).into(),
 					receiver: as_h160_32bytes(Keyring::TrancheInvestor(2)),
 					amount: AMOUNT,
@@ -302,6 +316,15 @@ fn transfer_tranche_tokens_domain_to_local<T: Runtime>() {
 	utils::prepare_hold_tt_domain::<T>(&mut env);
 
 	env.state_mut(|evm| {
+		assert!(
+			Decoder::<Balance>::decode(&evm.view(
+				Keyring::TrancheInvestor(1),
+				names::POOL_A_T_1,
+				"balanceOf",
+				Some(&[Token::Address(Keyring::TrancheInvestor(1).into())]),
+			)) >= AMOUNT,
+			"Insufficient POOL_A_T_1 funds by TrancheInvestor(1)"
+		);
 		evm.call(
 			Keyring::TrancheInvestor(1),
 			Default::default(),
@@ -321,6 +344,8 @@ fn transfer_tranche_tokens_domain_to_local<T: Runtime>() {
 			Some(&[
 				Token::Uint(POOL_A.into()),
 				Token::FixedBytes(pool_a_tranche_1_id::<T>().into()),
+				Token::Uint(DOMAIN_CENTRIFUGE.into()),
+				Token::Uint(CENTRIFUGE_CHAIN_ID.into()),
 				Token::FixedBytes(Keyring::TrancheInvestor(2).id().to_raw_vec()),
 				Token::Uint(AMOUNT.into()),
 			]),
